add a ProcessFunction that resembles the Window API
This AggregatingWindowWithProcessFunction implements parts of the Window API
with support for sliding and tumbling event-time windows (no session windows!).
Trigger (and purge) flexibility is limited and there are no evictors or
processing-time support.
Where possible, Window API interfaces where used but at some points, we needed
to define new ones to cope with the different setup inside a ProcessFunction.